webpack-config: Configure minifier to preserve translator comments (#21667)
This drops use of calypso-build's terser configuration in favor of one
that preserves the translator comments (and doesn't keep the various
licensing comments inline). It also simplifies the calculation of
terser's `ecma` option to match how terser actually uses it (e.g. terser
doesn't do anything different for "2015" versus "2019").

As for css-minimizer-webpack-plugin, this updates it from version 1.3.0
(used by calypso-build) to 3.1.1, which brings a new minification
technique that sorts the declarations for better compression.

This also fixes some code in Jetpack where the minifier loses translator
comments.

Fixes #16549