Use monorepo `validate-es` script to validate Webpack builds (#21729)
This may help in the future to ensure that the Webpack, Babel, and/or
Terser configs don't wind up out of sync with the browserslist
requirements.

Also fix a bug in webpack-config:
`browserslist.findConfig()` will probably return
"extends @wordpress/browserslist-config", not the actual list of
browsers. We need to pass it through `browserslist()` for that.