Revert and fix "Revert "Assets: do not use the new method yet (#21760)"" (#21763)
The new `Assets::register_script` wrapper worked great. In fact, a bit
too good: it picked up a style dependency that was missing from the
editor JS, which broke the Widget Editor and the Site Editor because
that dependency is specifically not supposed to be loaded there
(see #20357).

This reverts #21760 and adds a hack around the way that #20357 made
things break with the new wrapper.