jetpack: Avoid generating unused JS for static-site-generator assets
We use a Webpack static-site-generator plugin to build some HTML files.
This has the side effect of building redundant JS and (in one case) CSS
assets that we don't use.

Let's create a small Webpack plugin to allow us to suppress these unused
assets to reduce the size of the plugin. This may also be of use in the
future if/when we switch the SASS builds to use Webpack instead of gulp.