Remove special cases for calling changelogger on packages/changelogger
We have special cases in a number of our tools to account for the fact
that Composer doesn't create a `vendor/bin/changelogger` symlink in
packages/changelogger's vendor dir.

Instead, let's use post-install and post-upgrade scripts to ensure that
the link exists, so we don't have to special case it everywhere else.