webpack-config: Add a `CssRule` config
We had everything repeating basically the same boilerplate for building
a CSS rule. We can avoid that by making a `CssRule` that accepts the
additional loaders needed by some projects.

Also, it turns out that the `cache-loader` we copied from
calypso-build's CSS handling rules is deprecated in Webpack 5. So let's
drop it.