jetpack: Use Webpack to build modules JS
Currently the build uses an old version of Terser via gulp-minify to
simply minify the various modules. Let's switch that over to Webpack
where we have Babel for transpilation, a newer version of Terser, and so
on.

This did necessitate a few changes to some of the existing modules:
* Adjust a few bundled libraries' comments to preserve licenses that
  were being preserved previously.
* Adjust `modules/likes/post-count.js` and `post-count-jetpack.js` to
  preserve the `wpPostLikeCount` global that it seems like was there
  intentionally.
* Adjust `modules/widgets/google-translate/google-translate.js` so the
  global function is still global.

A followup will then rename these back from ".min.js" and use
`Assets::register_script()` to register the scripts.

There's also now an opportunity here for someone to update the modules
to modern JS syntax (now that we have Babel in the pipeline), to pull in
libraries like swiper via package.json instead of bundling them manually
(now that we're using an actual bundler), and maybe even to use wp.i18n.
I'll leave all that for someone else to do.