Debug Helper Plugin: add a Sync Data Settings Tester page (#21786)
* Debug Helper Plugin: add a Sync Data Settings Tester page

It would be useful to have an easy way to observe the Sync data settings provided by the Sync data filters. Let's add a Sync Data Settings Tester page to the Debug Helper plugin. The output of the `apply_filters` call for each Sync data filter will be displayed on this page.