assets: Have `register_script()` take a textdomain
We're probably going to want to have Assets do something to interact
with the upcoming Webpack lazy-load i18n plugin when it registeres a
script using wp-i18n.

Even if that turns out not to be the case, this will still help make
sure that `wp_set_script_translations` gets called correctly when
needed.