Sharing: ask search engines to not index pages with query string.
Fixes #21836

Jetpack's unofficial (icon, icon+text, text) sharing buttons currently work like so:
- When you append a query string with the name of a sharing service to a post URL, we redirect to the social service in question, with information the service can use (e.g. the post title as a tweet message).
- That query string may look like `?share=twitter`
- A page with such query string is not directly accessible since you are immediately redirected to the social network.
- WP Core implements `canonical` rel links in the head of each post page, so if you were looking at such a page's source code before the redirect happens, you would see that the canonical link for that page is in fact without the query string.
- Our HTML markup for the sharing buttons includes a `nofollow` rel attribute in the links, so as to indicate to search engines that the link should not be followed.

Yet that does not seem to be enough, as Google Search Console still seems to complain about those links in some scenarios, as mentioned in the issue above.

I am honestly not sure the changes introduced in this commit will really help, but I do not think it will hurt to try. WP's Robots API allows us to add meta tags to pages with that query string so we add one more indicator to search engines that the pages should not be indexed.