Jetpack: Start an initial full sync if the site is connected when Jetpack is activated.
An initial full sync is started when the site is registered. If the site is already connected
via another plugin when Jetpack is activated, a Sync process won't be started until a change is
made on the site that triggers a Sync. So, let's start an initial full sync during plugin activation
if the site is already connected.