Composer Packages: change the package type to jetpack-library
Strings in packages that are installed in the 'vendor' directory are not translated. Therefore,
the Jetpack Composer packages should be installed in a different directory. The new jetpack-composer-plugin
package will install packages with the jetpack-library type in the 'jetpack_vendor' directory. Update the type
for each of the Jetpack Composer packages to jetpack-library.