tools: Fix changelogger-release.sh logic
When changelogger-release.sh calls check-intra-monorepo-deps.sh to
update projects individually, we need to not have
check-intra-monorepo-deps.sh try to update pnpm-lock.yaml because the
dependencies may not be in a consistent state until it has finished
updating all projects.

So when check-intra-monorepo-deps.sh is called with a list of projects
to update, rely on the caller (changelogger-release.sh) to do the pnpm
lock file update.