Fix lots of i18n errors
For the most part this is fixing minifier breakage where code like
```js
flag ? __( 'foo', 'domain' ) : __( 'bar', 'domain' )
```
gets transformed into
```js
__( flag ? 'foo', 'bar', 'domain' )
```
which breaks the detection of the two strings as needing translation.

This also cleans up a few places where the domain was omitted, and adds
some context to a few translations.