Do not display the card when the user is not connected yet
In the other cards, we display an invitation to connect the user to WordPress.com when it is not connected yet, since the upgrade is not possible without that user connection.
For now, let's hide the card entirely.

In a follow-up PR, we can opt in to display the card, but with a different message when the user is not connected.