Reconfigure eslint to pull from `.gitignore` and per-dir `.eslintignore`
Unfortunately we have to do it in a somewhat hacky manner: in eslintrc
we run a function that reads all the ignore files and dumps the results
into eslint's `ignorePatterns` config.

Then we have to do it all over again for any sub-eslintrc that does
`root: true` or that extends any eslintrc that has the hack. Inheritance
is ok, but apparently we don't do that all that often (sigh).