Pay with PayPal: switch to using a unique script tag
Fixes #22125

This should avoid issues with other plugins that may use a similar script tag for their own asset. This was a bit too generic.

I've opted to leave the other tag name (for the script loaded directly from PayPal), since a generic name may be a good fit there.