[not verified] Donations: Don't display on front-end if stripe disconnected
When adding/editing a document with a Donations block that doesn't have
stripe connected it displays a prominent message saying that the block
will be hidden to visitors until stripe is connected. In actual fact it
displays a partial, broken block.

This change does as the message says, and displays nothing on the site
front-end to visitors, unless stripe is connected.

Fixes #17643