Donations block: Reuse stored plan IDs (#22239)
here are some cases where the stored plan IDs in the attributes of the Donations block are erased if they are edited by a user who is not the plan owner. To solve that, these changes introduce a safety check to only edit the stored plan IDs if the new ones have a value.