i18n-loader-webpack-plugin: Move the actual downloading out of the runtime
Instead of being provided state that it uses to build the URL and fetch
it, it now expects to be provided a loader that will do all the work for
it.

This makes the runtime smaller and allows for more flexibility in the
future if we need to change how the fetching works.