Admin Page: don't show plugin cards when you can't install plugins (#22258)
* Admin Page: don't show plugin cards when you can't install plugins

This adds a new check to see if one can install plugins. This is especially important, where an admin with the manage_options cap may not have the capability to install plugins. There, only superadmins have that capability.

* Switch to checking for manage_plugins instead

That seems more relevant to our needs in any case, and solves the issue highlighted here:
https://github.com/Automattic/jetpack/pull/22258#pullrequestreview-846829714

Co-authored-by: Samiff <samiff@users.noreply.github.com>