Simple payments: Hide paypal when invalid
When the paypal block was misconfigured it was still rendered on the
front-end. When people tried to use it they would receive an opaque
error message from paypal.

To resolve this, the paypal block will not be rendered on the front end
if it's misconfigured. There are plenty of messages displayed in the
editor if it is misconfigured so post-editors should be able to see any
problems before they happen. This makes it more consistent with the
donations block which does not display anything if not connected to
stripe.

Fixes #17641