[not verified] Payments block: Add transform from buttons
Allow transforming core/buttons blocks in Gutenberg to
jetpack/recurring-payments. This is only allowed if the sole child block
of core/buttons is a core/button, which requires gutenberg >= 11.5.0 to
detect. The restriction is because core/buttons can contain many
core/button instances, converting each to a jetpack/recurring-payments
block isn't possible. Transformations can only transform to the type
they're defined in and a jetpack/recurring-payments with multiple
buttons doesn't match with the way it works.

Fixes #17249