cli: Add dependency analysis helper and command
This is the existing logic from tools/find-project-deps.php and
tools/get-build-order.php, rewritten in JS. It also has a little more
flexibility as far as processing only a subset of projects.

One difference is in the unused composer.json `.extra.dependencies`
array: it will want that to instead be an object holding
`.extra.dependencies.test` and `.extra.dependencies.build`.

Changes to make use of this in place of tools/find-project-deps.php and
tools/get-build-order.php will be done as a followup.