Remove old dependency analysis scripts, use the new cli command
Other than needing `pnpm install`, the new cli command should be a
direct replacement for the old scripts. In a few cases we can take
advantage of the new `jetpack dependencies build-order` not filtering
out the monorepo by default.

Also update the `.extra.dependencies` in i18n-check-webpack-plugin, and
the relevant docs.