Avoid local eslint errors from e2e dirs
We have the e2e test directories in separate pnpm workspaces because
they require large downloads for playwright that are not normally needed
in local development. But that can cause problems when trying to run
eslint as the plugins referenced from eslintrc are not available.

To avoid that, let's adjust the eslint configurations in the e2e dirs to
detect this situation and avoid trying to lint at all.