[not verified] Premium content: Don't offer impossible transforms
In Gutenberg >= 11.1.0 don't offer transforms to premium contents from
blocks that can't support it, i.e. those that are already premium
content blocks or have premium content blocks in their lineage.

In older versions of gutenberg the transformations will continue to be
offered but will continue to be no-ops.

Fixes #22423