cli: Fix friendly error for missing modules
The CLI used to print a friendly error message when modules were
missing, but that was broken in #22395. This restores it.

Also, this adds output of the original error message for cases where the
problem isn't fixed by `pnpm install`. And since I'm doing it anyway, it
also tries to bold the final line on other errors.