Premium content: Improve editing consistency
In the editor, viewing the content for people who don't have a
subscription is sometimes referred to as 'Visitor view' and sometimes as
'Logged out view'. This change makes the interface more consistent by using 'Guest View'.

Swap order of visitor/subscriber buttons in the editor to match the
order of the added content.

Fixes #22425