cli: Build in parallel!
When building multiple projects, `jetpack build` will now build them in
parallel when possible. A `--concurrency` option is added to limit the
maximum number of concurrent project builds.

We can't quite use this in the GitHub Action yet, as for that we first
need to move all the before-build and after-build code into the build
action somehow. That's left for a future PR.