Widget Visibility: switch to using shared js package utility
In #22527, I copied over a utility function that we used in the block bundle to use in the Widget visibility bundle. However, starting in #22733, that utility function is now part of an external js package that we can use in the Widget visiblity bundle. Let's do that and save some duplication!