cli: Remove addition of changelogger in `--for-mirrors`
This was originally added (with an incorrect slug) because the mirroring
invokes changelogger in order to update the changelog.

But since the mirroring also forces building of dependencies, and it
only runs changelogger if the dependencies call for it, we don't
actually need to do that since changelogger will already have been
called for that way.