Publicize: Sync Publicize_UI with WPCOM
In order to land the changes in #22511 we need to make sure that the
files are in sync with WPCOM. The coding standards have changed and so
when we sync'd this UI we had to add a note for translators.

This brings the Jetpack version in line with that change.