Photon: ensure fallback when 3rd-parties empty array of sources
Fixes #22692

Photon hooks into the `wp_calculate_image_srcset` filter to add Photon domain and parameters to srcset image URLs generated for each image in a post. However, other plugins can also hook into that same filter, and some empty the array of images available for srcset. Let's support that use-case by checking not only for an array of images, but for an array that is not empty.