pr-is-up-to-date: Allow checking multiple tags
In our monorepo, we don't want releases of one project to make PRs
against another out of date. The solution is to have one tag per project
(plus the global tag for situations when CI infrastructure changes) and
only check the PR against the relevant project.

This needs two changes in the action:

1. When called on PR, allow specifying multiple tags.
2. When called on tag push, allow specifying the paths the pushed tag
   affects.