actions: Maintain per-project pr-update-to tags (#22860)
We're going to start using per-project pr-update-to tags to avoid e.g. a
release of Jetpack from invalidating PRs touching only Boost. Part of
that is that we'll need to create all those tags; that will be done
manually. But then we also need to keep those tags up to date as things
are released, which is what this PR is for.

Actually using the tags will come later. For now this also updates the
global pr-update-to tag whenever a per-project tag is updated.