status: Cache return values
Some of the Status methods are called many times during a request. Cache
them in process so as to avoid many calls to underlying methods that may
be somewhat expensive.

This also fixes `is_local_site()` and `is_staging_site()` to only call
`site_url()` once on an uncached call.

Fixes #22667