autoloader: Fix coverage generation for upstream change (#23062)
sebastianbergmann/php-code-coverage@0f89e2c renamed some internal
classes, one of which we were using to munge line numbers for the
autoloader coverage report.

Unfortunately there doesn't really seem to be any cleaner way to do
this. At the least we can read the class name with `get_class()`
instead of hard coding it.

Also, since it came up while trying to test this locally, set
`COMPOSER_HOME` when trying to run old versions of composer that don't
recognize new-format GitHub OAuth tokens so it doesn't complain about
the one that's probably in the local config.