actions: Direct CI Slack notifications to different channels
We've decided to split the notifications, with master going to one
channel and release branches going to another.

There are two ways this could have been done: have a step determine the
channel as an output to feed to the send-message step, or have
send-message steps per channel with `if` used to enable the appropriate
steps. I decided to go with the latter, since it's more straightforward
that way if at some point in the future we decide messages might want to
go to more than one channel (e.g. say Boost wanted notifications about
their release branch to go to their channel too).