Remove isLarge, it is not used anymore.
This should get us rid of warnings such as:

```
Warning: React does not recognize the `isLarge` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `islarge` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
```