repo-gardening: Consider both old and new names of moved files as modified (#23354)
We've long had a problem where the "add changelog entries" check was
incorrectly detecting that no change entry was added. Looking into this,
I find that it seems to always be release PRs, where GitHub is detecting
the removed change files as being renamed into a different project.

The fix is to have our `getFiles()` function record both the old and new
names of renamed files so they can be properly considered.