Premium content: Toolbar improvements
 * Use drop-down for view switcher on smaller viewports
 * Use BlockControls groups rather than ToolBarGroup
 * Use a ToolbarDropdownMenu rather than a DropdownMenu for plans

== View switcher ==
Move the view-switcher to a drop-down rather than two separate buttons
on smaller viewports so that it takes up less space. This means that 
mobile users can now see more of the menu without scrolling. On 
medium and upwards viewports, continue displaying as two separate
buttons to reduce the amount of clicking needed to switch between
guest and subscriber views.

The view switcher is now also a separate component

== ToolbarGroup removal ==
Using BlockControls groups rather than ToolbarGroups has fixed an issue
where the three-dots 'more' menu would float behind other menu items
when the menubar is fixed, rather than being able to be scrolled to.

== ToolbarDropdownMenu ==

Using a ToolbarDropdownMenu rather than a DropdownMenu for
selecting a plan has fixed an issue where the BlockControls would
disappear when resizing the viewport to be smaller than medium.


Fixes #22745, #23337