js-tools: Simplify eslint react configs (#23360)
We can save a bit of effort in various projects' eslint configs if we
let the react eslint config be used to extend the standard config
instead of having to `root: true` it. And we can clean up the monorepo
root dir slightly by moving the extendable eslint configs into
`tools/js-tools/`.

This also speeds up the linting slightly. Not a whole lot, maybe 5%.

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>