Pay with Paypal block: Fixed inconsistent currency formatting (#23368)
* Fixed inconsistent currency formatting in Pay with Paypal block.

Move the siteLocale registration inside Jetpack_Gutenberg, right after the 'jetpack-blocks-editor' is registered, since it's been already used by two blocks: Opentable and Pay with Paypal.

siteLocale property was not properly set in the Customizer because the wp_add_inline_script() call was execute before the "jetpack-blocks-editor" script was registered. When the script is registering, all the associated scripts were unregistered manually. Because of this, the currency format generated in the Widgets was different from the one generated in Customizer, thus triggering a block invalidation issue.

* Added siteLocale property in the wp_localize_script of jetpack-blocks-editor.

Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>