Update snapshots
Our Webpack config assigns integer keys to the various packages, based
on a hash of the path (which, for pnpm, includes the version number
too). It looks like one of the updated babel modules hashes to a
two-digit number rather than a three-digit number, throwing off the line
numbers in the snapshots.