beta: Set `Update URI` in the plugin header
WordPress 5.8 [added an "Update URI" plugin header][1], to make it clearer
when a plugin is not supposed to be hosted in the WordPress.org
repository. This helps prevent them from being accidentally submitted
there, and prevents accidental updates if someone somehow gets a plugin
registered there under the name.

Since the Jetpack Beta plugin cannot be hosted there (its whole point
goes against their rule 8, AFAICT), let's set that header.

WP 5.8 also introduced a new filter that could be used for the Beta
plugin's self-updates, but changing to use that is left for a future PR.
We still need to use the existing method for managed plugins anyway.

[1]: https://wp.me/p2AvED-nam