linting: Fix monorepo version refs lint
It seems likely to have been broken since #22433 added a call to the
Jetpack CLI without there being a `pnpm install` first. Add that to the
action.

Also update the script to detect this sort of error and handle it
better, instead of trying to lint a project with the improbably name of
"? Install the following package: jetpack@latest? (Y/n) ‣ ", and pass
the `--no-install` flag to `pnpx` in a few more places for good measure.