Add template lock to premium-content buttons
Currently users can add extra payment blocks to premium-content buttons,
If the button has the same planId as the existing one, then there's no
point adding it - it'll unlock the same content, and have the same
payment terms. If the button has a different planId as the existing one,
then it won't unlock the content it looks to be associated with.