Change style used for JS dependencies (#23653)
It's annoying to have to update all the dependencies of monorepo JS
packages for every patch-level version bump. Let's make it work like the
PHP dependencies, where the dependencies are only updated on minor and
major bumps.

The new scheme looks like `workspace:* || ^1.2`. Within the monorepo the
`workspace:*` part is operative and the `^1.2` is ignored by pnpm but
will still be checked by CI.

When we publish we can strip the `workspace:* || ` bit leaving only the
`^1.2` for external users. But I'll save that change for a followup PR.