pre-push hook: Don't fail if no tty is available
PR #23359 added code to have the pre-push hook run `jetpack changelog add`,
but this code fails if the push is happening via an IDE where a tty is
not available.

The fix is in two parts: First, don't redirect stdin from /dev/tty in
.husky/pre-push if it's going to fail. Second, don't try to run
`jetpack` in pre-push-hook.js if stdin isn't a tty.