tools: Fix use of jq `inside( array )` (#23702)
jq's `$arr1 | inside( $arr2 )`, when `$arr1` and `$arr2` are string
arrays, doesn't test for whether the elements of the input $arr1 are
members of $arr2. It tests for whether all strings in `$arr1` are
substrings of some string in `$arr2`, which is generally not the
behavior we want.

The same goes for `contains()`, but we're not using it on arrays
anywhere.

Instead, we can use `index()` to check for whether a string is in an
array.

Then fix Boost deps that the broken check didn't detect in #23482.