Google Fonts: mark feature as Beta and remove from dashboard.
Follow-up from #23295

The feature is still under development, and we'll need to make some more changes before it's ready for prime-time. Let's make that clear in the plugin by marking the module as beta and removing the toggle from the dashboard to avoid setting wrong expectations.

Internal reference: p1HpG7-flD-p2#comment-53417