tools: Fix node version check in check-development-environment.sh
The minimum version set in package.json is like `^14.18.3 || ^16.13.2`.
We weren't handling that correctly, we were comparing with a minimum
"version" of "14.18.3 || ^16.13.2" rather than anything well-formed.

This changes the check to extract the constraint matching the current
major version of node (i.e. "14.18.3" or "16.13.2"), falling back to the
recommended major if someone has a completely unsupported major version
(e.g. 15 or 17) installed, to use as the minimum version in the check.

This still produces slightly misleading results in that if someone is on
node 12 or 15 it won't tell them that 14 is an option, but that's
probably ok since we only have 14 in there because Gutenberg Mobile is
still on that version for their build (cf. #20990).