build: Avoid races when writing composer.json
Every once in a while the parallel build fails because one process is
writing composer.json at the same moment another process tries to read
it. The reader sees an empty or incomplete file, and composer chokes.

We can avoid these races by writing the file to a temporary name then
renaming it, as a file rename should be an atomic operation.