jetpack: Minifiy non-sass css with webpack
This allows them to take care of the better optimization in the newer
toolchain, and lets us work towards reducing dependencies by
consolidating on a single frontend build tool.

This also drops a few minified files that had been checked in to the
repo, and comments out an attempt at using a sass variable in a css
file that the toolchain now complains about.