cli: Avoid including ever-changing "reference" in vendor/ files
Since we switched to parallel builds, the "reference" fields that
Composer calculates by hashing the composer.json and path-repo options
have been changing with every build, due to nondeterministic content of
those path-repo options.

This makes it deterministic: include only the versions for the
dependencies of the package being built, and include them in
alphabetical order.